moved utils file into control directory
